# src/main/resources/META-INF/openapi/auth/auth-paths.yaml
openapi: 3.1.0
info:
  contact:
    email: support@example.com
    name: Support
  description: Authentication API for Figma Plugin
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Authentication API
  version: 0.0.1

tags:
  - name: Figma Authentication
    description: Figma OAuth authentication endpoints
  - name: GitHub Authentication
    description: GitHub OAuth authentication endpoints
  - name: Auth Resource
    description: Authentication and token management endpoints

paths:
  /auth/access-token/refresh:
    post:
      summary: Refresh access token
      description: Generates a new access token using a refresh token
      parameters:
        - name: userId
          in: query
          schema:
            type: string
          required: true
          description: User ID for token refresh
          example: "user_123456789"
        - name: refresh_token
          in: cookie
          schema:
            anyOf:
              - $ref: "../common/common-schemas.yaml#/Cookie"
              - type: "null"
          description: Refresh token cookie
      responses:
        "200":
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: "./auth-schemas.yaml#/RefreshTokenResponse"
              examples:
                successful_refresh:
                  $ref: "./auth-examples.yaml#/RefreshTokenResponse/successful_refresh"
                admin_refresh:
                  $ref: "./auth-examples.yaml#/RefreshTokenResponse/admin_refresh"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              examples:
                invalid_token:
                  $ref: "./auth-examples.yaml#/ErrorResponses/invalid_refresh_token"
      tags:
        - Auth Resource

  /auth/refresh-token:
    get:
      summary: Get refresh token
      description: Generates and returns a refresh token cookie for the authenticated user
      responses:
        "200":
          description: Refresh token generated successfully and set as HTTP-only cookie
          content:
            application/json:
              examples:
                success:
                  $ref: "./auth-examples.yaml#/RefreshTokenGeneration/success"
        "401":
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              examples:
                unauthorized:
                  $ref: "./auth-examples.yaml#/ErrorResponses/unauthorized"
        "500":
          description: Internal server error
          content:
            application/json:
              examples:
                server_error:
                  $ref: "./auth-examples.yaml#/ErrorResponses/server_error"
      security:
        - JWT: [ ]
      tags:
        - Auth Resource

  /auth/figma/login:
    get:
      summary: Initiate Figma OAuth login
      description: Starts the Figma OAuth authentication flow by returning the authorization URL and read token
      tags:
        - Figma Authentication
      responses:
        "200":
          description: Login URL generated successfully
          content:
            application/json:
              schema:
                $ref: "./auth-schemas.yaml#/OAuthInitResponse"
              examples:
                figma_login:
                  $ref: "./auth-examples.yaml#/OAuthInitResponse/figma_login"
        "500":
          description: Internal server error - failed to generate login URL
          content:
            application/json:
              examples:
                login_generation_error:
                  $ref: "./auth-examples.yaml#/ErrorResponses/login_generation_error"

  /auth/figma/callback:
    get:
      summary: Handle Figma OAuth callback
      description: Processes the OAuth callback from Figma with authorization code and state
      tags:
        - Figma Authentication
      parameters:
        - description: Authorization code from Figma
          required: true
          name: code
          in: query
          schema:
            type: string
          example: "figma_auth_code_abc123def456"
        - description: State parameter for CSRF protection
          required: true
          name: state
          in: query
          schema:
            type: string
          example: "csrf_state_xyz789"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              examples:
                successful_callback:
                  $ref: "./auth-examples.yaml#/CallbackResponses/figma_success"
        "400":
          description: Bad request - invalid code or state
          content:
            application/json:
              examples:
                invalid_code:
                  $ref: "./auth-examples.yaml#/ErrorResponses/invalid_callback_code"
                invalid_state:
                  $ref: "./auth-examples.yaml#/ErrorResponses/invalid_callback_state"

  /auth/figma/access-token:
    get:
      summary: Get Figma access token
      description: Retrieves the access token for authenticated user to access Figma API
      tags:
        - Figma Authentication
      responses:
        "200":
          description: Access token retrieved successfully
          content:
            application/json:
              schema:
                $ref: "./auth-schemas.yaml#/AccessTokenResponse"
              examples:
                figma_token:
                  $ref: "./auth-examples.yaml#/AccessTokenResponse/figma_token"
        "401":
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              examples:
                unauthorized:
                  $ref: "./auth-examples.yaml#/ErrorResponses/unauthorized"
        "408":
          description: Request timeout - access token not available
          content:
            application/json:
              examples:
                token_timeout:
                  $ref: "./auth-examples.yaml#/ErrorResponses/token_timeout"
      security:
        - JWT: [ ]

  /auth/github/login:
    get:
      summary: Initiate GitHub OAuth login
      description: Starts the GitHub OAuth authentication flow by returning the authorization URL and read token
      tags:
        - GitHub Authentication
      responses:
        "200":
          description: Login URL generated successfully
          content:
            application/json:
              schema:
                $ref: "./auth-schemas.yaml#/OAuthInitResponse"
              examples:
                github_login:
                  $ref: "./auth-examples.yaml#/OAuthInitResponse/github_login"
        "500":
          description: Internal server error - failed to generate login URL
          content:
            application/json:
              examples:
                login_generation_error:
                  $ref: "./auth-examples.yaml#/ErrorResponses/login_generation_error"

  /auth/github/callback:
    get:
      summary: Handle GitHub OAuth callback
      description: Processes the OAuth callback from GitHub with authorization code and state
      tags:
        - GitHub Authentication
      parameters:
        - description: Authorization code from GitHub
          required: true
          name: code
          in: query
          schema:
            type: string
          example: "github_auth_code_def456ghi789"
        - description: State parameter for CSRF protection
          required: true
          name: state
          in: query
          schema:
            type: string
          example: "csrf_state_abc123"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              examples:
                successful_callback:
                  $ref: "./auth-examples.yaml#/CallbackResponses/github_success"
        "400":
          description: Bad request - invalid code or state
          content:
            application/json:
              examples:
                invalid_code:
                  $ref: "./auth-examples.yaml#/ErrorResponses/invalid_callback_code"
                invalid_state:
                  $ref: "./auth-examples.yaml#/ErrorResponses/invalid_callback_state"

  /auth/github/access-token:
    get:
      summary: Get GitHub access token
      description: Retrieves the access token for authenticated user to access GitHub API
      tags:
        - GitHub Authentication
      responses:
        "200":
          description: Access token retrieved successfully
          content:
            application/json:
              schema:
                $ref: "./auth-schemas.yaml#/AccessTokenResponse"
              examples:
                github_token:
                  $ref: "./auth-examples.yaml#/AccessTokenResponse/github_token"
        "401":
          description: Unauthorized - invalid or missing authentication
          content:
            application/json:
              examples:
                unauthorized:
                  $ref: "./auth-examples.yaml#/ErrorResponses/unauthorized"
        "408":
          description: Request timeout - access token not available
          content:
            application/json:
              examples:
                token_timeout:
                  $ref: "./auth-examples.yaml#/ErrorResponses/token_timeout"
      security:
        - JWT: [ ]